+++ internal/erasure/memstats_test.go
@@ -1,0 +1,10 @@
+package erasure
+
+import "runtime"
+
+type testingMemStats = runtime.MemStats
+
+func readMem(m *runtime.MemStats) {
+	runtime.GC()
+	runtime.ReadMemStats(m)
+}